refactor events module
