+++ Innovation_WebApp/admin.py
@@ -1,10 +1,8 @@
 from django.contrib import admin
-from .models import Events,SubscribedUsers,EventRegistration,CommunityProfile,CommunitySession,Community,CommunityMember,Social_media
+from .models import SubscribedUsers,CommunityProfile,CommunitySession,Community,CommunityMember,Social_media
 
 # Register your models here.
-admin.site.register(Events)
 admin.site.register(SubscribedUsers)
-admin.site.register(EventRegistration)
 admin.site.register(CommunityProfile)
 admin.site.register(CommunitySession)
 admin.site.register(Community)
