+++ events/Email.py
@@ -2,6 +2,8 @@ from django.core.mail import EmailMultiAlternatives
 from django.template.loader import render_to_string
 from django.utils.html import strip_tags
 
+
+
 def send_ticket_email(registration):
     ticket_details = {
         'ticket_number': str(registration.ticket_number),
@@ -15,7 +17,6 @@ def send_ticket_email(registration):
     subject = f'Event Ticket: {registration.event.name}'
     html_message = render_to_string('email/send_email.html', {'ticket_details': ticket_details})
     plain_message = strip_tags(html_message)
-    
 
     email = EmailMultiAlternatives(
         subject,
@@ -27,13 +28,14 @@ def send_ticket_email(registration):
     email.attach_alternative(html_message, "text/html")
     email.send(fail_silently=False)
 
-def send_the_otp_email(user,otp):
+
+def send_the_otp_email(user, otp):
     otp_details = {
-        'user':user.first_name,
-        'otp':otp.otp_code
+        'user': user.first_name,
+        'otp': otp.otp_code
     }
     subject = 'Verify Your Email Address'
-    html_message = render_to_string('OTP/otp.html',{'otp_details':otp_details})
+    html_message = render_to_string('OTP/otp.html', {'otp_details': otp_details})
     plain_message = strip_tags(html_message)
 
     email = EmailMultiAlternatives(
@@ -43,6 +45,5 @@ def send_the_otp_email(user,otp):
 
         [user.email]
     )
-    email.attach_alternative(html_message,"text/html")
+    email.attach_alternative(html_message, "text/html")
     email.send(fail_silently=False)
-
